3863 Moving PipelineVisitor
