+++ src/blocks/PipelineVisitor.test.ts
@@ -17,7 +17,7 @@
 
 import { DocumentRenderer } from "@/blocks/renderers/document";
 import ForEach from "@/blocks/transformers/controlFlow/ForEach";
-import { BlockConfig } from "@/blocks/types";
+import { BlockPosition, BlockConfig } from "@/blocks/types";
 import { createNewElement } from "@/components/documentBuilder/createNewElement";
 import { Schema } from "@/core";
 import { PIPELINE_BLOCKS_FIELD_NAME } from "@/pageEditor/consts";
@@ -28,7 +28,6 @@ import {
   typedBlockFactory,
 } from "@/testUtils/factories";
 import { toExpression } from "@/testUtils/testHelpers";
-import { AbsolutePosition } from "./analysisTypes";
 import PipelineVisitor, { VisitBlockExtra } from "./PipelineVisitor";
 
 jest.mock("@/blocks/registry", () => ({
@@ -45,7 +44,7 @@ test("should invoke the callback for the pipeline bricks", async () => {
   const visitBlock = jest.fn();
   class Visitor extends PipelineVisitor {
     override async visitBlock(
-      position: AbsolutePosition,
+      position: BlockPosition,
       blockConfig: BlockConfig,
       extra: VisitBlockExtra
     ) {
@@ -90,7 +89,7 @@ test("should invoke the callback for the sub pipeline bricks", async () => {
 
   class Visitor extends PipelineVisitor {
     override async visitBlock(
-      position: AbsolutePosition,
+      position: BlockPosition,
       blockConfig: BlockConfig,
       extra: VisitBlockExtra
     ) {
@@ -147,7 +146,7 @@ test("should invoke the callback for the Document button pipeline", async () => 
 
   class Visitor extends PipelineVisitor {
     override async visitBlock(
-      position: AbsolutePosition,
+      position: BlockPosition,
       blockConfig: BlockConfig,
       extra: VisitBlockExtra
     ) {
